+++ src/extension.ts
@@ -3,6 +3,7 @@ import { registerCommands } from "./commands";
 import { registerCommentController } from "./comments";
 import { registerFileSystemProvider } from "./fileSystem";
 import { log } from "./logger";
+import { initializeProtocolHander } from "./protocolHandler";
 import { store } from "./store";
 import { initializeAuth } from "./store/auth";
 import { initializeStorage } from "./store/storage";
@@ -11,7 +12,9 @@ import { registerTreeProvider } from "./tree";
 export async function activate(context: vscode.ExtensionContext) {
   try {
     log.setLoggingChannel(vscode.window.createOutputChannel('GistPad'));
+
     initializeAuth();
+    initializeProtocolHander();
 
     registerCommands(context);
     registerFileSystemProvider(store);